teach54

+++ teaching_threads/par_map.c
@@ -39,8 +39,8 @@ double *par_map(double *list, size_t list_len, mapper map_func, size_t num_threa
 
 	double *res = (double *)malloc(list_len * sizeof(double));
 
-	int block_size = list_len / num_threads;
-	printf("%d\n", block_size);
+	size_t block_size = list_len / num_threads;
+	printf("%zu\n", block_size);
 
 	pthread_t ** threads = malloc(num_threads * sizeof(pthread_t *));
 	blazeit ** arguments = malloc(num_threads * sizeof(blazeit));
